build background image and form
used flexbox to devide the body into two parts
one for background image which take up 35% base
using flex: 1 1 35% and the 65% rest belong to
form accessories.
in the form again another flexbox is used to
align the header, form elelemts, and submit
vertically.
the third flexbox is used to align the label and
input vertically as well.

[Ticket: 1]