Fix viewport problem
the .form-bottom elements crossed the base
viewport and caused the scrolling come into
effect. to remove this behavior two actions
have been done.
1. removed the margin-top on the container-header
it was 150px, it is then cancelled. Instead of
that the header-container is set to display:flex
and align-item: end. also 20% padding right has
applied to push the text to the left side of the
page.

2. set the .form-details element to display:flex
flex-direction: column and justify-content: end
when we put justify content to the end the video
and the submit part both will stick to the bottom
of the page and cancelled scrolling effect.
[Ticket: 9]